requirements.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
